chore: Remove lightning and sub channel remnants
