+++ crates/xxi-node/src/dlc/contract_details.rs
