+++ crates/xxi-node/src/node/connection.rs
@@ -1,8 +1,5 @@
 use crate::bitcoin_conversion::to_secp_pk_29;
-use crate::bitcoin_conversion::to_secp_pk_30;
 use crate::networking;
-use crate::node::event::NodeEvent;
-use crate::node::event::NodeEventHandler;
 use crate::node::Node;
 use crate::node::NodeInfo;
 use crate::node::Storage;
@@ -13,71 +10,8 @@ use anyhow::Context;
 use anyhow::Result;
 use bitcoin::secp256k1::PublicKey;
 use futures::Future;
-use lightning::events::OnionMessageProvider;
-use lightning::ln::features::InitFeatures;
-use lightning::ln::features::NodeFeatures;
-use lightning::ln::msgs;
-use lightning::ln::msgs::OnionMessage;
-use lightning::ln::msgs::OnionMessageHandler;
 use std::pin::Pin;
-use std::sync::Arc;
 use std::time::Duration;
-
-pub struct TenTenOneOnionMessageHandler {
-    handler: Arc<NodeEventHandler>,
-}
-
-impl TenTenOneOnionMessageHandler {
-    pub fn new(handler: Arc<NodeEventHandler>) -> Self {
-        TenTenOneOnionMessageHandler { handler }
-    }
-}
-
-/// Copied from the IgnoringMessageHandler
-impl OnionMessageProvider for TenTenOneOnionMessageHandler {
-    fn next_onion_message_for_peer(
-        &self,
-        _peer_node_id: bitcoin_old::secp256k1::PublicKey,
-    ) -> Option<OnionMessage> {
-        None
-    }
-}
-
-/// Copied primarily from the IgnoringMessageHandler. Using the peer_connected hook to get notified
-/// once a peer successfully connected. (This also includes that the Init Message has been processed
-/// and the connection is ready to use).
-impl OnionMessageHandler for TenTenOneOnionMessageHandler {
-    fn handle_onion_message(
-        &self,
-        _their_node_id: &bitcoin_old::secp256k1::PublicKey,
-        _msg: &OnionMessage,
-    ) {
-    }
-    fn peer_connected(
-        &self,
-        their_node_id: &bitcoin_old::secp256k1::PublicKey,
-        _init: &msgs::Init,
-        inbound: bool,
-    ) -> Result<(), ()> {
-        tracing::info!(%their_node_id, inbound, "Peer connected!");
-
-        self.handler.publish(NodeEvent::Connected {
-            peer: to_secp_pk_30(*their_node_id),
-        });
-
-        Ok(())
-    }
-    fn peer_disconnected(&self, _their_node_id: &bitcoin_old::secp256k1::PublicKey) {}
-    fn provided_node_features(&self) -> NodeFeatures {
-        NodeFeatures::empty()
-    }
-    fn provided_init_features(
-        &self,
-        _their_node_id: &bitcoin_old::secp256k1::PublicKey,
-    ) -> InitFeatures {
-        InitFeatures::empty()
-    }
-}
 
 impl<D: BdkStorage, S: TenTenOneStorage + 'static, N: Storage + Sync + Send + 'static>
     Node<D, S, N>
