+++ crates/xxi-node/src/node/dlc_channel.rs
@@ -805,9 +805,9 @@ impl<D: BdkStorage, S: TenTenOneStorage + 'static, N: LnDlcStorage + Sync + Send
 /// Use this instead of [`MessageHandler`]'s `send_message` which only enqueues the message.
 ///
 /// [`MessageHandler`]: dlc_messages::message_handler::MessageHandler
-pub fn send_dlc_message<D: BdkStorage, S: TenTenOneStorage + 'static, N: LnDlcStorage>(
+pub fn send_dlc_message<D: BdkStorage>(
     dlc_message_handler: &TenTenOneMessageHandler,
-    peer_manager: &PeerManager<D, S, N>,
+    peer_manager: &PeerManager<D>,
     node_id: PublicKey,
     msg: TenTenOneMessage,
 ) {
