+++ crates/xxi-node/src/tests/mod.rs
@@ -1,7 +1,5 @@
 use crate::bitcoin_conversion::to_secp_pk_29;
 use crate::bitcoin_conversion::to_xonly_pk_29;
-use crate::config::app_config;
-use crate::config::coordinator_config;
 use crate::node::dlc_channel::send_dlc_message;
 use crate::node::event::NodeEvent;
 use crate::node::event::NodeEventHandler;
@@ -31,10 +29,8 @@ use dlc_manager::payout_curve::PayoutPoint;
 use dlc_manager::payout_curve::PolynomialPayoutCurvePiece;
 use dlc_manager::payout_curve::RoundingInterval;
 use dlc_manager::payout_curve::RoundingIntervals;
-use dlc_manager::subchannel::SubChannelState;
 use dlc_manager::ReferenceId;
 use futures::Future;
-use lightning::util::config::UserConfig;
 use rand::distributions::Alphanumeric;
 use rand::thread_rng;
 use rand::Rng;
@@ -86,7 +82,6 @@ impl Node<on_chain_wallet::InMemoryStorage, TenTenOneInMemoryStorage, InMemorySt
     fn start_test_app(name: &str) -> Result<(Arc<Self>, RunningNode)> {
         Self::start_test(
             name,
-            app_config(),
             ELECTRS_ORIGIN.to_string(),
             OracleInfo {
                 endpoint: ORACLE_ORIGIN.to_string(),
@@ -112,7 +107,6 @@ impl Node<on_chain_wallet::InMemoryStorage, TenTenOneInMemoryStorage, InMemorySt
     ) -> Result<(Arc<Self>, RunningNode)> {
         Self::start_test(
             name,
-            coordinator_config(),
             ELECTRS_ORIGIN.to_string(),
             OracleInfo {
                 endpoint: ORACLE_ORIGIN.to_string(),
@@ -126,7 +120,6 @@ impl Node<on_chain_wallet::InMemoryStorage, TenTenOneInMemoryStorage, InMemorySt
     #[allow(clippy::too_many_arguments)]
     fn start_test(
         name: &str,
-        ldk_config: UserConfig,
         electrs_origin: String,
         oracle: OracleInfo,
         node_storage: Arc<InMemoryStore>,
@@ -150,7 +143,6 @@ impl Node<on_chain_wallet::InMemoryStorage, TenTenOneInMemoryStorage, InMemorySt
         let (dlc_event_sender, dlc_event_receiver) = mpsc::channel::<DlcChannelEvent>();
         let event_handler = Arc::new(NodeEventHandler::new());
         let node = Node::new(
-            ldk_config,
             name,
             Network::Regtest,
             data_dir.as_path(),
@@ -308,37 +300,6 @@ fn random_tmp_dir() -> PathBuf {
     tmp
 }
 
-#[allow(dead_code)]
-fn log_channel_id(
-    node: &Node<on_chain_wallet::InMemoryStorage, TenTenOneInMemoryStorage, InMemoryStore>,
-    index: usize,
-    pair: &str,
-) {
-    let details = match node.channel_manager.list_channels().get(index) {
-        Some(details) => details.clone(),
-        None => {
-            tracing::info!(%index, %pair, "No channel");
-            return;
-        }
-    };
-
-    let channel_id = hex::encode(details.channel_id.0);
-    let short_channel_id = details.short_channel_id;
-    let is_ready = details.is_channel_ready;
-    let is_usable = details.is_usable;
-    let inbound = details.inbound_capacity_msat / 1000;
-    let outbound = details.outbound_capacity_msat / 1000;
-    tracing::info!(
-        channel_id,
-        short_channel_id,
-        is_ready,
-        is_usable,
-        inbound,
-        outbound,
-        "{pair}"
-    );
-}
-
 async fn wait_until<P, T, F>(timeout: Duration, predicate_fn: P) -> Result<T>
 where
     P: Fn() -> F,
@@ -353,46 +314,6 @@ where
         }
     })
     .await?
-}
-
-#[derive(PartialEq, Debug)]
-enum SubChannelStateName {
-    Offered,
-    Accepted,
-    Confirmed,
-    Finalized,
-    Signed,
-    Closing,
-    OnChainClosed,
-    CounterOnChainClosed,
-    CloseOffered,
-    CloseAccepted,
-    CloseConfirmed,
-    OffChainClosed,
-    ClosedPunished,
-    Rejected,
-}
-
-impl From<&SubChannelState> for SubChannelStateName {
-    fn from(value: &SubChannelState) -> Self {
-        use SubChannelState::*;
-        match value {
-            Offered(_) => SubChannelStateName::Offered,
-            Accepted(_) => SubChannelStateName::Accepted,
-            Confirmed(_) => SubChannelStateName::Confirmed,
-            Finalized(_) => SubChannelStateName::Finalized,
-            Signed(_) => SubChannelStateName::Signed,
-            Closing(_) => SubChannelStateName::Closing,
-            OnChainClosed => SubChannelStateName::OnChainClosed,
-            CounterOnChainClosed => SubChannelStateName::CounterOnChainClosed,
-            CloseOffered(_) => SubChannelStateName::CloseOffered,
-            CloseAccepted(_) => SubChannelStateName::CloseAccepted,
-            CloseConfirmed(_) => SubChannelStateName::CloseConfirmed,
-            OffChainClosed => SubChannelStateName::OffChainClosed,
-            ClosedPunished(_) => SubChannelStateName::ClosedPunished,
-            Rejected => SubChannelStateName::Rejected,
-        }
-    }
 }
 
 fn xxi_node_settings_coordinator() -> XXINodeSettings {
